Reorganize operator definitions
