+++ src/opdefs/reroll.ts
